requirements.txt
dnspython==2.2.1
pynetbox==6.6.2
PyYAML==6.0
